Removed Cells and Views which will be added in a different gem + fixed gem folder structure and namespacing
