+++ lib/tyrant/change_password.rb
@@ -1,0 +1,2 @@
+require "tyrant/change_password/contract/change_password"
+require "tyrant/change_password/operation/change_password"
