--- lib/tyrant/view/change_password.slim
@@ -1,26 +1,0 @@
-.row
-  h1 Change Password
-
-  = trb_form_for(model, url_for(controller: "users", action: "change_password"), builder: :bootstrap4, id: :change_password) do |f|
-    .row
-      .col-sm-2
-      .col-sm-8
-        = f.input :email, placeholder: "Email"
-      .col-sm-2
-    .row
-      .col-sm-2
-      .col-sm-8
-        = f.input :password, placeholder: "Password", type: "password"
-      .col-sm-2
-    .row
-      .col-sm-2
-      .col-sm-4
-        = f.input :new_password, placeholder: "New Password", type: "password"
-      .col-sm-4
-        = f.input :confirm_new_password, placeholder: "Confirm New Password", type: "password"
-      .col-sm-2
-    .row
-      .col-sm-2
-      .col-sm-8
-        = f.submit(value: "Change Password")
-      .col-sm-2
